+++ changelog.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## [1.13.7] - 2026-9-9 Boxscore Bug Busting
+## [1.13.7] - 2026-9-10 Brazen Bot Bugs
 
 ### New Features
 - A guild linked Tournament can now send reminder messages through discord. Setup under the manage guild page by guild moderators
 - /timestamp command split out from /schedule set. The timestamp command simply displays the time with the timestamp for easy copying
+- /lfg now has a title field that is used as the embed title, thread title and game nickname
 
 ### Improvements
 - Better banner message for prefilled matches
@@ -20,6 +21,8 @@
 - Alert for forge submitted factions didn't link to submission queue
 - Closed polls no longer say the time was rejected if there was no roster to compare to
 - Editing a match no longer kills open schedule polls
+- Games recorded from an LFG thread will no longer have blank discarded captain fields
+- Games recorded from a Match thread or LFG thread will no longer seat players in an order if there is no established order.
 
 ## [1.13.6] - 2026-9-8 Box Score API & Availability